Harder better faster stronger `new-or-deleted-file`
